apps/py-chat/requirements.txt
neo4j
fastapi
uvicorn
python-dotenv
sentence_transformers